Add result widget and reset button
This adds a result widget that displays the result of the quiz.

In this I modified the questions back to become similar to questions
used in the tutorial, and added a result widget that showed a page
with personality results, and a reset button to reset the app.

Once again this code is following the tutorial and we haven't started
building the actual app yet.